hotfix unique + filter functions
