+++ ui/templates/done.html
@@ -33,7 +33,7 @@
     {% csrf_token %}
 
   
-    <label for="feature-list">Choose a fruit:</label>
+    <label for="feature-list">Choose a configuration element:</label>
     <br>
     <select class="feature-list-class" name="feature-list" id="feature-list"> 
       {% for m in collapsable %}
